+++ R/utils.R
@@ -37,7 +37,7 @@ stan_fns_as_string <- function(files, include) {
 
 #' Load a package example
 #'
-#' Loads examples of nowcasts produce using example scripts. Used to streamline
+#' Loads examples of nowcasts produced using example scripts. Used to streamline
 #' examples, in package tests and to enable users to explore package
 #' functionality without needing to install `cmdstanr`.
 #'
@@ -430,8 +430,9 @@ utils::globalVariables(
     "run_time", "cum_prop_reported", "..by_with_group_id",
     "reference_missing", "prop_missing", "day", "posteriors",
     "formula", ".id", "n", ".confirm_avail", "prediction", "true_value",
-    "person", "id", "latest", "num_reference_date", "num_report_date",
-    "rep_mod", "ref_mod", "count", "reference_date_mod", "report_date_mod",
-    "timestep", ".observed", "lookup", "max_obs_delay"
+    "person", "id", "latest", "type", "below_coverage", "num_reference_date",
+    "num_report_date", "rep_mod", "ref_mod", "count", "reference_date_mod",
+    "report_date_mod", "timestep", ".observed", "lookup", "max_obs_delay",
+    "coverage"
   )
 )